Moved Twitter example to the where the other examples are.
